added post and output by post category

+++ restaurant/models.py
@@ -17,7 +17,7 @@ class RangeMenu(models.Model):
     title = models.CharField(verbose_name='Названия', max_length = 150)
     image = models.ImageField(verbose_name='Изображение', upload_to='images_menu/')
     description = models.TextField(verbose_name='Описание')
-    price = models.SmallIntegerField(verbose_name='Цена', default='')
+    price = models.DecimalField(verbose_name='Цена', max_digits=5, decimal_places=2, default='')
     category_menu = models.ForeignKey(CategoryMenu, verbose_name='Меню ресторана', on_delete=models.CASCADE)
     slug = models.SlugField(max_length = 150, unique=True)
